Extract EventBus trait and move the Akka based implementation into impl.util
